5056 Remove query-lambdav2 module
