+++ .github/config/chunks.yaml
@@ -85,6 +85,5 @@ chunks:
       - query/query-runner
       - query/query-runnerv2
       - query/query-lambda
-      - query/query-lambdav2
       - athena
       - trino
