+++ .github/workflows/chunk-query.yaml
@@ -13,7 +13,6 @@ on:
       - 'java/query/query-runner/**'
       - 'java/query/query-runnerv2/**'
       - 'java/query/query-lambda/**'
-      - 'java/query/query-lambdav2/**'
       - 'java/athena/**'
       - 'java/trino/**'
       - 'java/partitions/splitter/**'
