+++ java/distribution/src/main/assembly/zip.xml
@@ -145,7 +145,7 @@
             <destName>lambda-splitter-${project.version}.jar</destName>
         </file>
         <file>
-            <source>${project.basedir}/../query/query-lambdav2/target/query-lambdav2-${project.version}-utility.jar</source>
+            <source>${project.basedir}/../query/query-lambda/target/query-lambda-${project.version}-utility.jar</source>
             <outputDirectory>scripts/jars</outputDirectory>
             <destName>query-${project.version}.jar</destName>
         </file>
