+++ java/query/pom.xml
@@ -31,6 +31,5 @@
         <module>query-runner</module>
         <module>query-lambda</module>
         <module>query-runnerv2</module>
-        <module>query-lambdav2</module>
     </modules>
 </project>
