+++ java/query/query-lambda/src/main/java/sleeper/query/lambda/WarmQueryExecutorLambda.java
@@ -15,21 +15,18 @@
  */
 package sleeper.query.lambda;
 
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
 import com.amazonaws.services.lambda.runtime.events.ScheduledEvent;
-import com.amazonaws.services.s3.AmazonS3;
-import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.amazonaws.services.sqs.AmazonSQS;
-import com.amazonaws.services.sqs.AmazonSQSClientBuilder;
-import com.amazonaws.services.sqs.model.SendMessageRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import software.amazon.awssdk.services.dynamodb.DynamoDbClient;
+import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.sqs.SqsClient;
+import software.amazon.awssdk.services.sqs.model.SendMessageRequest;
 
-import sleeper.configuration.properties.S3InstanceProperties;
-import sleeper.configuration.properties.S3TableProperties;
+import sleeper.configurationv2.properties.S3InstanceProperties;
+import sleeper.configurationv2.properties.S3TableProperties;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.range.Range;
 import sleeper.core.range.Region;
@@ -44,7 +41,7 @@ import sleeper.query.core.model.Query;
 import sleeper.query.core.model.QueryProcessingConfig;
 import sleeper.query.core.model.QuerySerDe;
 import sleeper.query.core.output.ResultsOutputConstants;
-import sleeper.query.runner.tracker.DynamoDBQueryTracker;
+import sleeper.query.runnerv2.tracker.DynamoDBQueryTracker;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -54,7 +51,7 @@ import java.util.UUID;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.CONFIG_BUCKET;
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.QUERY_QUEUE_URL;
 import static sleeper.core.properties.table.TableProperty.TABLE_NAME;
-import static sleeper.query.runner.output.NoResultsOutput.NO_RESULTS_OUTPUT;
+import static sleeper.query.runnerv2.output.NoResultsOutput.NO_RESULTS_OUTPUT;
 
 /*
  * A Lambda that is triggered when an {@link ScheduledEvent} is received. A processor executes, creates a new
@@ -67,17 +64,17 @@ public class WarmQueryExecutorLambda implements RequestHandler<ScheduledEvent, V
     private static final Logger LOGGER = LoggerFactory.getLogger(WarmQueryExecutorLambda.class);
 
     private final InstanceProperties instanceProperties;
-    private final AmazonSQS sqsClient;
-    private final AmazonS3 s3Client;
-    private final AmazonDynamoDB dynamoClient;
+    private final SqsClient sqsClient;
+    private final S3Client s3Client;
+    private final DynamoDbClient dynamoClient;
     private final DynamoDBQueryTracker queryTracker;
 
     public WarmQueryExecutorLambda() throws ObjectFactoryException {
-        this(AmazonS3ClientBuilder.defaultClient(), AmazonSQSClientBuilder.defaultClient(),
-                AmazonDynamoDBClientBuilder.defaultClient(), System.getenv(CONFIG_BUCKET.toEnvironmentVariable()));
+        this(S3Client.create(), SqsClient.create(),
+                DynamoDbClient.create(), System.getenv(CONFIG_BUCKET.toEnvironmentVariable()));
     }
 
-    public WarmQueryExecutorLambda(AmazonS3 s3Client, AmazonSQS sqsClient, AmazonDynamoDB dynamoClient, String configBucket) throws ObjectFactoryException {
+    public WarmQueryExecutorLambda(S3Client s3Client, SqsClient sqsClient, DynamoDbClient dynamoClient, String configBucket) throws ObjectFactoryException {
         this.s3Client = s3Client;
         this.sqsClient = sqsClient;
         this.dynamoClient = dynamoClient;
@@ -130,7 +127,10 @@ public class WarmQueryExecutorLambda implements RequestHandler<ScheduledEvent, V
                             .build();
 
                     LOGGER.info("Query to be sent: " + querySerDe.toJson(query));
-                    SendMessageRequest message = new SendMessageRequest(instanceProperties.get(QUERY_QUEUE_URL), querySerDe.toJson(query));
+                    SendMessageRequest message = SendMessageRequest.builder()
+                            .queueUrl(instanceProperties.get(QUERY_QUEUE_URL))
+                            .messageBody(querySerDe.toJson(query))
+                            .build();
                     LOGGER.debug("Message: {}", message);
                     sqsClient.sendMessage(message);
                 });
