--- java/query/query-lambdav2/src/test/java/sleeper/query/lambda/WarmQueryExecutorLambdaIT.java
@@ -1,156 +1,0 @@
-/*
- * Copyright 2022-2025 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package sleeper.query.lambda;
-
-import com.amazonaws.services.lambda.runtime.events.ScheduledEvent;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest;
-import software.amazon.awssdk.services.sqs.model.ReceiveMessageResponse;
-
-import sleeper.configurationv2.properties.S3InstanceProperties;
-import sleeper.configurationv2.properties.S3TableProperties;
-import sleeper.configurationv2.table.index.DynamoDBTableIndexCreator;
-import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.core.properties.table.TableProperties;
-import sleeper.core.range.Range;
-import sleeper.core.range.Region;
-import sleeper.core.schema.Field;
-import sleeper.core.schema.Schema;
-import sleeper.core.schema.type.StringType;
-import sleeper.core.statestore.StateStore;
-import sleeper.core.util.ObjectFactoryException;
-import sleeper.localstack.test.LocalStackTestBase;
-import sleeper.query.core.model.Query;
-import sleeper.query.core.model.QueryProcessingConfig;
-import sleeper.query.core.model.QuerySerDe;
-import sleeper.query.core.output.ResultsOutputConstants;
-import sleeper.query.runnerv2.tracker.DynamoDBQueryTrackerCreator;
-import sleeper.statestorev2.StateStoreFactory;
-import sleeper.statestorev2.transactionlog.TransactionLogStateStoreCreator;
-
-import java.io.IOException;
-import java.nio.file.Path;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-
-import static java.nio.file.Files.createTempDirectory;
-import static org.assertj.core.api.Assertions.assertThat;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.CONFIG_BUCKET;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.DATA_BUCKET;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.LEAF_PARTITION_QUERY_QUEUE_URL;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.QUERY_QUEUE_URL;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.QUERY_RESULTS_BUCKET;
-import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.QUERY_RESULTS_QUEUE_URL;
-import static sleeper.core.properties.instance.CommonProperty.FILE_SYSTEM;
-import static sleeper.core.properties.instance.TableDefaultProperty.DEFAULT_INGEST_PARTITION_FILE_WRITER_TYPE;
-import static sleeper.core.properties.table.TableProperty.TABLE_NAME;
-import static sleeper.core.properties.testutils.InstancePropertiesTestHelper.createTestInstanceProperties;
-import static sleeper.core.properties.testutils.TablePropertiesTestHelper.createTestTableProperties;
-import static sleeper.core.statestore.testutils.StateStoreUpdatesWrapper.update;
-import static sleeper.query.runnerv2.output.NoResultsOutput.NO_RESULTS_OUTPUT;
-
-public class WarmQueryExecutorLambdaIT extends LocalStackTestBase {
-
-    private QuerySerDe querySerDe;
-
-    @TempDir
-    public Path tempDir;
-    private InstanceProperties instanceProperties;
-    protected static WarmQueryExecutorLambda lambda;
-
-    @BeforeEach
-    void setUp() throws IOException, ObjectFactoryException {
-        String dataDir = createTempDirectory(tempDir, null).toString();
-        createInstanceProperties(dataDir);
-        lambda = new WarmQueryExecutorLambda(s3ClientV2, sqsClientV2, dynamoClientV2, instanceProperties.get(CONFIG_BUCKET));
-    }
-
-    @Test
-    public void shouldCreateAQueryWithKeyTypeOfString() throws Exception {
-        // Given
-        Schema schema = getStringKeySchema();
-        TableProperties tableProperties = createTestTableProperties(instanceProperties, schema);
-        createTable(tableProperties);
-        querySerDe = new QuerySerDe(schema);
-
-        // When
-        lambda.handleRequest(new ScheduledEvent(), null);
-
-        // Then
-        ReceiveMessageResponse result = sqsClientV2.receiveMessage(ReceiveMessageRequest.builder()
-                .queueUrl(instanceProperties.get(QUERY_QUEUE_URL))
-                .build());
-
-        assertThat(result.messages()).hasSize(1);
-
-        Query query = querySerDe.fromJson(result.messages().get(0).body());
-        Query expected = buildExpectedQuery(query.getQueryId(), tableProperties.get(TABLE_NAME), schema,
-                new Field("test-key", new StringType()), "a");
-
-        assertThat(query).isEqualTo(expected);
-    }
-
-    private Schema getStringKeySchema() {
-        return Schema.builder()
-                .rowKeyFields(new Field("test-key", new StringType()))
-                .sortKeyFields(new Field("test-sort", new StringType()))
-                .valueFields(new Field("test-value", new StringType()))
-                .build();
-    }
-
-    private Query buildExpectedQuery(String id, String tableName, Schema schema, Field rowKey, Object value) {
-        Region region = new Region(List.of(new Range.RangeFactory(schema)
-                .createExactRange(rowKey, value)));
-        return Query.builder()
-                .queryId(id)
-                .tableName(tableName)
-                .regions(List.of(region))
-                .processingConfig(QueryProcessingConfig.builder()
-                        .resultsPublisherConfig(Map.of(ResultsOutputConstants.DESTINATION, NO_RESULTS_OUTPUT))
-                        .statusReportDestinations(Collections.emptyList())
-                        .build())
-                .build();
-    }
-
-    private void createInstanceProperties(String dir) throws IOException {
-        instanceProperties = createTestInstanceProperties();
-        instanceProperties.set(DEFAULT_INGEST_PARTITION_FILE_WRITER_TYPE, "direct");
-        instanceProperties.set(FILE_SYSTEM, "file://");
-        instanceProperties.set(DATA_BUCKET, dir);
-        instanceProperties.set(CONFIG_BUCKET, "testing");
-        instanceProperties.set(QUERY_QUEUE_URL, createSqsQueueGetUrl());
-        instanceProperties.set(LEAF_PARTITION_QUERY_QUEUE_URL, createSqsQueueGetUrl());
-        instanceProperties.set(QUERY_RESULTS_QUEUE_URL, createSqsQueueGetUrl());
-        instanceProperties.set(QUERY_RESULTS_BUCKET, dir + "/query-results");
-
-        createBucket(instanceProperties.get(CONFIG_BUCKET));
-        S3InstanceProperties.saveToS3(s3ClientV2, instanceProperties);
-
-        new DynamoDBQueryTrackerCreator(instanceProperties, dynamoClientV2).create();
-        DynamoDBTableIndexCreator.create(dynamoClientV2, instanceProperties);
-        new TransactionLogStateStoreCreator(instanceProperties, dynamoClientV2).create();
-    }
-
-    private void createTable(TableProperties tableProperties) {
-        S3TableProperties.createStore(instanceProperties, s3ClientV2, dynamoClientV2).save(tableProperties);
-        StateStore stateStore = new StateStoreFactory(instanceProperties, s3ClientV2, dynamoClientV2)
-                .getStateStore(tableProperties);
-        update(stateStore).initialise(tableProperties.getSchema());
-    }
-}
